Support GOARCH !amd64

+++ atime_darwin.go
@@ -8,5 +8,5 @@ import (
 
 func fileInfoAccessTime(fi os.FileInfo) time.Time {
 	ts := fi.Sys().(*syscall.Stat_t).Atimespec
-	return time.Unix(ts.Sec, ts.Nsec)
+	return time.Unix(int64(ts.Sec), int64(ts.Nsec))
 }
